requirements.txt
Django==4.2.6
django-crispy-forms==2.0
crispy-bootstrap5==0.7
requests==2.28.1
django-jazzmin==2.6.0 
django-simple-history==3.4.0